make playwright init beautiful
