+++ auth/rod.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"time"
 
-	"github.com/charmbracelet/huh/spinner"
 	"github.com/rusq/slackauth"
 
 	"github.com/rusq/slackdump/v3/auth/auth_ui"
@@ -155,18 +154,9 @@ func headlessFlow(ctx context.Context, cl *slackauth.Client, workspace string, u
 		return sp, fmt.Errorf("password cannot be empty")
 	}
 
-	sctx, stopSpinner := context.WithCancel(ctx)
-	defer stopSpinner()
-	go func() {
-		_ = spinner.New().
-			Type(spinner.Dots).
-			Title("Logging in to Slack, it will take 25-40 seconds").
-			Context(sctx).
-			Run()
-	}()
-
+	stopSpinnerFn := pleaseWait(ctx, "Logging in to Slack, it will take 25-40 seconds")
 	var loginErr error
-	sp.Token, sp.Cookie, loginErr = cl.Headless(ctx, username, password, stopSpinner)
+	sp.Token, sp.Cookie, loginErr = cl.Headless(ctx, username, password, stopSpinnerFn)
 	if loginErr != nil {
 		return sp, loginErr
 	}
